+++ src/vector/scene/scene_fill.cpp
@@ -115,7 +115,7 @@ static void fill_gradient(VectorState &State, const TClipRectangle<double> &Boun
    const agg::trans_affine &Transform, double ViewWidth, double ViewHeight, extVectorGradient &Gradient,
    GRADIENT_TABLE *Table, agg::renderer_base<agg::pixfmt_psl> &RenderBase, agg::rasterizer_scanline_aa<> &Raster)
 {
-   constexpr LONG MAX_SPAN = 256;
+   constexpr int MAX_SPAN = 256;
    typedef agg::span_interpolator_linear<> interpolator_type;
    typedef agg::span_allocator<agg::rgba8> span_allocator_type;
    typedef agg::pod_auto_array<agg::rgba8, MAX_SPAN> color_array_type;
@@ -583,7 +583,7 @@ static void fill_pattern(VectorState &State, const TClipRectangle<double> &Bound
 
    transform *= Transform;
    transform.invert();
-   
+
    if (SampleMethod IS VSM::AUTO) {
       // Using anything more sophisticated than bicubic sampling for tiling is a CPU killer.
       // If the client requires a different method, they will need to set it explicitly.
